requirements.txt
dnspython
requests
pyyaml
rich
urllib3
